fix(podman): socket hang up (#33)
Stopping the container takes some time. If we do not kill the podman API
server first, the upper-layer application will mistakenly assume that
the API service is ready, leading to socket hang up issues during
subsequent operations.

Therefore, we need to kill the podman API server first and then stop the
container to avoid this problem.
In the future, we will use an event notification solution to completely
avoid this issue.